pci: make segment number 32-bit for devicetree support later
Segment/domain number obtained from devicetree is 32-bit. In addition,
current devicetree binding documents from Linux and Open Firmware do not
mention about its range.

It is time to change segment number type from u16 to u32 to prepare for
devicetree support. The u32 type also matches with SegmentNumber of
EFI_PCI_ROOT_BRIDGE_IO_PROTOCOL in UEFI specification.

However, ACPI system expects 16-bit segment number. The caller of
ACPI functions should check whether the segment number can be represented
in 16-bit before passing it to the functions.

+++ core/acpi.c
@@ -840,7 +840,7 @@ get_acpi_time_raw (u32 *r)
 }
 
 bool
-acpi_read_mcfg (uint n, u64 *base, u16 *seg_group, u8 *bus_start,
+acpi_read_mcfg (uint n, u64 *base, u32 *seg_group, u8 *bus_start,
 		u8 *bus_end)
 {
 	if (!saved_mcfg)
